refactor: RecommendationStatus 위치 변경
