+++ src/main/java/com/generic4/itda/domain/recommendation/constant/RecommendationAlgorithm.java
@@ -20,14 +20,4 @@ public enum RecommendationAlgorithm {
         this.code = code;
     }
 
-    @Getter
-    public enum RecommendationRunStatus {
-        PENDING("대기중"), RUNNING("실행중"), COMPUTED("계산 완료"), FAILED("실패");
-
-        private final String description;
-
-        RecommendationRunStatus(String description) {
-            this.description = description;
-        }
-    }
 }
